cephadm: move tracing class to a new file
