+++ src/cephadm/cephadmlib/daemons/__init__.py
@@ -1,3 +1,4 @@
 from .custom import CustomContainer
+from .tracing import Tracing
 
-__all__ = ['CustomContainer']
+__all__ = ['CustomContainer', 'Tracing']
